customized and redesigned the like button; added various title for client components
